For windows 11 Installer
The ei.cfg file is a configuration file used by Microsoft Windows installation media. It stands for "Edition ID Configuration" and is used to control which edition (version) of Windows is installed from the installation media.

When installing Windows from a DVD or USB drive, the installer looks for this file. If it finds the ei.cfg file, it reads the edition information from it. This allows system administrators and users to create customized installation media that installs a specific edition of Windows without having to manually select it during the installation process.

For instance, if you have a Windows installation ISO that contains multiple editions (such as Home, Pro, Enterprise), you can modify the ei.cfg file to specify which edition should be installed by default.

If the ei.cfg file is absent or doesn't contain edition information, the installer typically presents the user with a choice of editions to install.